removed macros from their own module
